backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
boto3==1.39.14
botocore==1.39.14
cryptography==46.0.3
email-validator==2.1.0
fastapi==0.116.1
idna==3.10
jmespath==1.0.1
mangum==0.19.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
s3transfer==0.13.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
